enforcer/tests/requirements.txt
coverage==4.5.4
google-api-python-client==1.7.11
google-cloud-resource-manager==0.29.2
google-cloud-storage==1.20.0
mock==3.0.5
pytest<5
pytest-cov==2.8.1